refactor: move interface/types and constants to other files
